jaruda_back/requirements.txt
asgiref==3.5.1
attrs==21.4.0
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
drf-spectacular==0.22.1
inflection==0.5.1
jsonschema==4.5.1
mysqlclient==2.1.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
celery>=5.0.5




